Adding high precision files
Adding a new rmd file which shows the results of changing the precision on the model threshold to >2187, rather than the precision threshold obtained from the experimental consultation.  This inevitably results in a much harsher rules of thumb decision tree and fewer species which we can model.  This rmd relies on the ev_trees_hi output uploaded today.